Fazendo os testes de remover produto, faltava a verificação se o mesmo está no banco... Adicionei esta funcionalidade no método removeProduto() para rodar os testes...

+++ src/java/ifpb/monteiro/ads/pd/DAO/ProdutoDAO.java
@@ -31,12 +31,17 @@ public class ProdutoDAO extends DAO<Produto> {
 
     @Override
     public void remove(Produto produto) throws HumQueCaroException {
+        Produto p = procura(produto.getCodigo());
         try {
-            abrirBanco();
-            getStmt().execute(
-                    "DELETE FROM produtos WHERE codigo = '" + produto.getCodigo()
-                    + "' ");
-            fecharBanco();
+            if (p != null) {
+                abrirBanco();
+                getStmt().execute(
+                        "DELETE FROM produtos WHERE codigo = '" + produto.getCodigo()
+                        + "' ");
+                fecharBanco();
+            } else {
+                throw new HumQueCaroException("Produto não encontrado");
+            }
         } catch (Exception e) {
             throw new HumQueCaroException("Erro no remove de Produto "
                     + e.getMessage());
